pyproject.toml
[project]
name = "mcp-kaggle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.5",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "kaggle>=2.0.0",
    "matplotlib>=3.10.8",
    "mcp[cli]>=1.26.0",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tavily-python>=0.7.21",
]
